use "unsigned long int" type instead of "int" for Cursor's field "cur", as calculations may cause integer overflow and turn it into negative on x86_64 platforms

+++ pykt/cursor.h
@@ -7,7 +7,7 @@
 
 typedef struct {
     PyObject_HEAD
-    int cur;
+    unsigned long int cur;
     DBObject *db;
 } CursorObject;
